pyproject.toml
[tool.poetry]
name = "generative-models"
version = "0.1.0"
description = "Implementation of generative models."
authors = ["PixelPenguin"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.5"
pandas = "^1.0.4"
Pillow = "^7.1.2"
PyYAML = "^5.3.1"
pytorch-lightning = "0.7.6"
scikit-learn = "^0.23.1"
tensorboard = "^2.2.2"
torch = "1.5.0"
torchvision = "0.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.3"
flake8 = "^3.8.2"
flake8-isort = "^3.0.0"
flake8-quotes = "^3.2.0"
jupyter = "^1.0.0"
jupyterlab = "^2.1.4"
pep8-naming = "^0.10.0"
tqdm = "^4.46.1"
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
